+++ StormLib/Services/Twitch/TwitchRecords.cs
@@ -1,5 +1,7 @@
+using System;
+
 namespace StormLib.Services.Twitch
 {
 	public record class TwitchDisplayName(string DisplayName);
-	public record class TwitchGame(int Id, string Name);
+	public record class TwitchGame(Int64 Id, string Name);
 }
